Adding mesh cleaning function to remove duplicated vertices
And degenerated triangles resulting for the duplicated vertives removal.

It uses `pymesh.remove_duplicated_vertices`.

Note that `pymesh.remove_degenerated_triangles` is finally not used
because it also modifiy the vertices (maybe only the order) but the
mapping is not returned in the informations, thus preventing to
appropriately update the mesh attributes.